feat: refacto jobOpportunity et queryValidators
