+++ server/src/services/application.service.ts
@@ -4,7 +4,7 @@ import { oleoduc, writeData } from "oleoduc"
 import { Application, LbaCompany, EmailBlacklist } from "../common/model/index.js"
 import { manageApiError } from "../common/utils/errorManager.js"
 import { decryptWithIV, encryptIdWithIV } from "../common/utils/encryptString.js"
-import { validateCaller } from "../service/queryValidators.js"
+import { validateCaller } from "./queryValidator.service.js"
 import { logger } from "../common/logger.js"
 import { prepareMessageForMail } from "../common/utils/fileUtils.js"
 import config from "../config.js"
