+++ server/src/services/jobOpportunity.service.types.ts
@@ -1,0 +1,15 @@
+export type TJobSearchQuery = {
+  romes?: string
+  rncp?: string
+  referer?: string
+  caller?: string
+  latitude?: string
+  longitude?: string
+  radius?: number
+  insee?: string
+  sources?: string
+  diploma?: string
+  opco?: string
+  opcoUrl?: string
+  useMock?: string
+}
